transformer, comment

+++ GraphVisualisierung2/src/main/java/vistra/core/graph/item/transformer/VertexValue.java
@@ -5,14 +5,19 @@ import org.apache.commons.collections15.Transformer;
 import vistra.core.graph.item.IVertexLayout;
 
 /**
+ * A vertex transformer: value.
+ * 
  * @author Roland Bruggmann (brugr9@bfh.ch)
  * 
  */
 public class VertexValue implements Transformer<IVertexLayout, String> {
 
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
 	public String transform(IVertexLayout vertex) {
-		return Double.toString(vertex.getValue());
+		return Integer.toString(vertex.getValue());
 	}
 
 }
